Split tree into more files
