+++ roots/TODOS.md
@@ -17,3 +17,11 @@ This is a temporary place to track TODOs for this project, until it's at the sta
 - [ ] Benchmarks
 - [ ] Ability to sequentially scan and reverse-scan transaction log.
 - [ ] Ability to find an entry in the transaction log, using a binary search
+- [ ] File versioning
+- [ ] Consider a WAL
+
+      In testing SQLite, they have a WAL which is used to batch changes into thir database. It slows down queries, but it allows inserts to be incredibly fast. Originally, I threw out this idea because it seemed like it might require too much memory usage.
+
+      But, I now realize we can contain the memory usage. In the chunk cacher, we can keep a separate list of cached chunks that are required to stay in memory until they're committed from the WAL. We can force this list to be a fixed size to keep memory usage down -- thus forcing a wait for the WAL to commit. Once the WAL has been committed, those chunks can be moved into the chunk cacher.
+
+      Why could this be faster? The WAL can contain just the bare minimum to re-execute the modification. This means less writes need to be flushed to disk to say that the transaction is completed. For flows that are doing small transactions, we could do fake transactions in memory that create temporary nodes that only live in the chunk cache. Then, when the WAL is committed, we do one large write to the database with all the changes. The only reason this might win is by minimizing the total number of node rewrites by chunking the btree operations into one batch.
