+++ roots/src/tree/state.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use tokio::sync::{Mutex, MutexGuard};
 
-use super::TreeRoot;
+use super::BTreeRoot;
 
 const UNINITIALIZED_SEQUENCE: u64 = 0;
 
@@ -24,7 +24,7 @@ impl<const MAX_ORDER: usize> State<MAX_ORDER> {
 #[derive(Debug, Default)]
 pub struct ActiveState<const MAX_ORDER: usize> {
     pub current_position: u64,
-    pub header: TreeRoot<MAX_ORDER>,
+    pub header: BTreeRoot<MAX_ORDER>,
 }
 
 impl<const MAX_ORDER: usize> ActiveState<MAX_ORDER> {
@@ -32,7 +32,7 @@ impl<const MAX_ORDER: usize> ActiveState<MAX_ORDER> {
         self.header.sequence != UNINITIALIZED_SEQUENCE
     }
 
-    pub fn initialize(&mut self, file_length: u64, header: TreeRoot<MAX_ORDER>) {
+    pub fn initialize(&mut self, file_length: u64, header: BTreeRoot<MAX_ORDER>) {
         assert_eq!(
             self.header.sequence, UNINITIALIZED_SEQUENCE,
             "state already initialized"
